ADded rest client, connection to Stooq, logs and error handling. Changed data types of quote's fields.

+++ bot/internal/domain/quote.go
@@ -5,9 +5,9 @@ import "time"
 type Quote struct {
 	Symbol string
 	Date   time.Time
-	Open   float32
-	High   float32
-	Low    float32
-	Close  float32
-	Volume uint64
+	Open   float64
+	High   float64
+	Low    float64
+	Close  float64
+	Volume int64
 }
